mailify-back/requirements.txt
Flask==2.2.3
flask-cors==3.1.1
openai==0.27.0
python-dotenv==1.0.0